test: 사다리 변환 출력에 대한 테스트 코드 작성
true 일때 '-----'로 변환하여 출력하고, false일때 '     '로 변환하여 출력하는 것에 도저히 믿음이 안가서 작성했습니다...
