+++ src/test/java/domain/LadderTest.java
@@ -33,22 +33,4 @@ public class LadderTest {
                         List.of(false, true, false),
                         List.of(true, false, true));
     }
-
-    static class TestLadderRowGenerator implements LadderRowGenerator {
-
-        private final List<List<Boolean>> ladder = List.of(
-                List.of(true, false, true),
-                List.of(false, true, false),
-                List.of(true, false, false),
-                List.of(false, true, false),
-                List.of(true, false, true)
-        );
-
-        private int index = 0;
-
-        @Override
-        public LadderRow generate(int userCount) {
-            return new LadderRow(ladder.get(index++));
-        }
-    }
 }
